phase-d(bun_runtime): h2_frame_parser: js_number f64 casts, create_utf8_for_js path, StringJsc to_js

+++ src/runtime/api/bun/h2_frame_parser.rs
@@ -2922,7 +2922,7 @@ impl H2FrameParser {
             // TODO: properly propagate exception upwards
             let end = content.end;
             self.read_buffer.reset();
-            self.dispatch_with_2_extra(JSH2FrameParser::Gc::onGoAway, JSValue::js_number(error_code), JSValue::js_number(self.last_stream_id as f64), chunk);
+            self.dispatch_with_2_extra(JSH2FrameParser::Gc::onGoAway, JSValue::js_number(error_code as f64), JSValue::js_number(self.last_stream_id as f64), chunk);
             return end;
         }
         data.len()
